requirements.txt
markdown
jinja2
weasyprint
fonttools[woff]
pillow
